Modify product commands to sync with job list.

+++ src/main/java/seedu/mycrm/logic/commands/products/DeleteProductCommand.java
@@ -2,8 +2,10 @@ package seedu.mycrm.logic.commands.products;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.mycrm.commons.core.Messages.MESSAGE_INVALID_PRODUCT_DISPLAYED_INDEX;
+import static seedu.mycrm.commons.core.Messages.MESSAGE_REMOVE_LINKED_PRODUCT;
 
 import java.util.List;
+import java.util.function.Predicate;
 
 import seedu.mycrm.commons.core.index.Index;
 import seedu.mycrm.logic.commands.Command;
@@ -11,6 +13,7 @@ import seedu.mycrm.logic.commands.CommandResult;
 import seedu.mycrm.logic.commands.CommandType;
 import seedu.mycrm.logic.commands.exceptions.CommandException;
 import seedu.mycrm.model.Model;
+import seedu.mycrm.model.job.Job;
 import seedu.mycrm.model.products.Product;
 
 public class DeleteProductCommand extends Command {
@@ -48,6 +51,19 @@ public class DeleteProductCommand extends Command {
         }
 
         Product productToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        // update job
+        Predicate<Job> jobPredicate = model.getJobPredicate() == null
+                ? model.PREDICATE_SHOW_ALL_INCOMPLETE_JOBS
+                : model.getJobPredicate();
+        model.updateFilteredJobList(Model.PREDICATE_SHOW_ALL_JOBS);
+        boolean isLinkedToJob = model.getFilteredJobList().stream()
+                .anyMatch(j -> j.getProduct() != null && j.getProduct().isSameProduct(productToDelete));
+        model.updateFilteredJobList(jobPredicate);
+        if (isLinkedToJob) {
+            throw new CommandException(MESSAGE_REMOVE_LINKED_PRODUCT);
+        }
+
         model.deleteProduct(productToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_PRODUCT_SUCCESS, productToDelete), COMMAND_TYPE);
     }
